requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiounittest==1.3.0
arrow==1.2.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
azure-core==1.24.2
azure-cosmos==3.2.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.7.0
binaryornot==0.4.4
botbuilder-ai==4.14.2
botbuilder-azure==4.14.2
botbuilder-core==4.14.2
botbuilder-dialogs==4.14.2
botbuilder-integration-aiohttp==4.14.2
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cookiecutter==1.7.0
cryptography==3.4.8
datatypes-date-time==1.0.0a2
datatypes-timex-expression==1.0.2a2
datedelta==1.4
emoji==1.7.0
et-xmlfile==1.1.0
Flask==2.1.2
frozenlist==1.3.0
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
grapheme==0.6.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.12.0
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==1.4.2
MarkupSafe==2.1.1
msal==1.6.0
msrest==0.6.19
msrestazure==0.6.4
multidict==6.0.2
multipledispatch==0.6.0
numpy==1.23.1
oauthlib==3.2.0
openai==0.22.0
openpyxl==3.0.10
pandas==1.4.3
pandas-stubs==1.4.3.220815
poyo==0.5.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
recognizers-date-time==1.0.0a1
recognizers-number==1.0.0a3
recognizers-number-with-unit==1.0.0a1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.8.1
tqdm==4.64.0
types-pytz==2022.2.1.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
whichcraft==0.6.1
wincertstore==0.2
yarl==1.4.2
zipp==3.8.0
